requirements.txt
# ─── Core Framework ───────────────────────────────────────────────────────────
streamlit>=1.33.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ─── OCR ──────────────────────────────────────────────────────────────────────
pytesseract>=0.3.10
easyocr>=1.7.1
opencv-python-headless>=4.9.0.80
Pillow>=10.3.0

# ─── PDF Conversion ───────────────────────────────────────────────────────────
pdf2image>=1.17.0
PyMuPDF>=1.24.0          # Fallback PDF converter (fitz)

# ─── NLP ──────────────────────────────────────────────────────────────────────
spacy>=3.7.4
langdetect>=1.0.9
transformers>=4.40.0     # Optional: transformer-based fallback extraction
# Note: download spaCy model after install:
#   python -m spacy download en_core_web_sm

# ─── Data Handling ────────────────────────────────────────────────────────────
pandas>=2.2.2
numpy>=1.26.4

# ─── Database ─────────────────────────────────────────────────────────────────
SQLAlchemy>=2.0.29

# ─── Search / Retrieval ───────────────────────────────────────────────────────
Whoosh>=2.7.4

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.1
python-multipart>=0.0.9  # FastAPI file uploads
httpx>=0.27.0            # Optional: async HTTP client

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-cov>=5.0.0
